refactor: extract team_orchestrator traits and types to astra-server-types
- Create DelegationExecutor, DelegationTracking, RunPersistence traits
  in astra-server-types for decoupling orchestrator from concrete types
- Move SubRunRecord, DelegationProgress to trait module (re-exported)
- Move ExecutionPhase, TeamExecutionStatus, OrchestratorConfig,
  ProgressCallback, and 7 helper functions to types module
- Move warn_persist! macro to astra-server-types (macro_export)
- Replace concrete types with trait objects in TeamExecutionOrchestrator
- Add trait impls for DelegationEngine, DelegationTracker, RunEngine
- Net: ~190 LOC moved from runtime to shared crate
