+++ rust/crates/astra-server-types/Cargo.toml
@@ -10,3 +10,5 @@ astra-services.workspace = true
 regex.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+async-trait.workspace = true
+tokio-util.workspace = true
